rename airtablUtils + add util to get login User
